Refactor: Remove existing tests for MachineASousComponent and MachineASousLogic, and add new comprehensive tests for FirebaseSendService and related components
- Deleted old test files for MachineASousComponent and MachineASousLogic.
- Created new test file for FirebaseSendService with detailed tests for its methods.
- Added new tests for MachineASousComponent, ensuring proper functionality and integration with Firebase.
- Enhanced test coverage for logic handling in MachineASousLogic, including error handling and data processing.
